+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - Support for iRODS 4.1.10.
 - iRODS setup code, split of from [test-with-baton](https://github.com/wtsi-hgi/test-with-baton) repository.
+- Helper for reading data object contents.
 
 ### Changed
 - iput helper no longer uses the project directory for temp files.